Resize canvas correctly when handheld devices are rotated (#15)
On handheld devices, such as tablets and phones, the
window.screen.height and window.screen.width don't actually change
(tested on an iPhone). This caused a third of the screen to be
unresponsive, for example, when rotating from portrait mode to landscape
mode.

The innerHeight and innerWidth properties need to be used instead as
they change with the device's rotation. On rotation, because the canvas
is resized, it will be cleared. However, this is already expected as
noted in the comment that handles the canvas resizing.